Commit 9 - 5/25/23
Release 1.0.0! Woo!

Creating this took much longer than expected. The GitHub website updated while I was working on this, which meant that some of the functionality broke. I used more fetch requests instead of window-related listening, which should improve the stability of the extension.

- Recognizes if a solution file exists or needs to be created and commits to or creates the file, respectively
- Recognizes if the solutions repo does not exist. If is is not found, it will attempt to create one
- All GitHub handling is done in the background; commits and file creations are no longer manual. Manual committing/creating may be re-added in the future as an option within settings.
- Added a demonstration video/GIF

Login pages could be removed later; they are unnecessary as long as the GitHub username is provided and valid